Vision fixed; comms is accurate (90%)
When accuracy fails, it's because the Arduino side registers the comms readings as starting at pos[1] or pos[2] and not pos[0], but the readings themselves are accurate. The Arduino code should display whether it's getting faulty readings, but it displays nothing -- so it's likely just a sorting logical error rather than a comms issue.